+++ satella/warnings.py
@@ -1,0 +1,7 @@
+__all__ = ['ExperimentalWarning']
+
+
+class ExperimentalWarning(FutureWarning):
+    """
+    This feature is experimental!
+    """